examples/requirements.txt
ofauth @ file://..
python-dotenv>=1.0.0
